refactor: 일부 dto domain 패키지로 이동
