+++ src/test/java/lotto/AmountTest.java
@@ -1,6 +1,6 @@
 package lotto;
 
-import lotto.dto.Amount;
+import lotto.domain.Amount;
 import lotto.exception.IllegalInputAmountException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
