+++ src/test/java/lotto/WinningCheckerTest.java
@@ -1,9 +1,9 @@
 package lotto;
 
+import lotto.domain.BonusNumber;
 import lotto.domain.Lotto;
+import lotto.domain.Lottos;
 import lotto.domain.WinningRank;
-import lotto.dto.BonusNumber;
-import lotto.dto.Lottos;
 import lotto.dto.WinningNumber;
 import lotto.dto.WinningStatistic;
 import lotto.service.helper.WinningChecker;
