+++ src/test/java/lotto/mock/MockInvalidManualLottoView.java
@@ -1,9 +1,9 @@
 package lotto.mock;
 
 import lotto.domain.Lotto;
-import lotto.dto.Lottos;
-import lotto.dto.ManualLottoNumbers;
-import lotto.dto.NumberOfManualLotto;
+import lotto.domain.Lottos;
+import lotto.domain.ManualLottoNumbers;
+import lotto.domain.NumberOfManualLotto;
 import lotto.dto.WinningStatistic;
 import lotto.view.View;
